first attempt to create separate app routes
